+++ src/Krystal/Templating/CHANGELOG.md
@@ -4,5 +4,6 @@ CHANGELOG
 1.3
 ---
 
+ * Added `wrap()` and `wrapMany()` methods
  * Added `StringTempalte` class with template() method to handle variables in strings
  * First version